# docs/swagger.yaml
definitions:
  handler.request:
    properties:
      active:
        type: boolean
      code:
        type: string
      color:
        type: string
      creationDate:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      published:
        type: boolean
      stock:
        type: integer
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/Support
  description: This API handles MELI products.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: Bootcamp - GO Web Module API
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: get products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Lists products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: store products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store products
      tags:
      - Products
  /products/{id}:
    delete:
      description: removes products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: product id to be removed
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Removes product based on given ID
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: updates name and price of products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: product id to be updated
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Updates name and price of product based on given ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: updates products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: product id to be updated
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Updates product based on given ID
      tags:
      - Products
swagger: "2.0"